Hackoholics/Flask-API/requirements.txt
Flask
flask-cors
python-dotenv
Pillow
pdf2image
google-generativeai